Major overhaul
Removed combine Excel as it was a basic version of merge Excel.  Moved code around in General Normalization. Deleted the code from CSV Normalize and Combine, copied in the code from General Normalization and removed all the if statements and graphing so that the code has one simple direct function. Added header comments to all of the files and added more comments where needed.